refactor(dashboard): share nav config between sidebar and mobile menu
