+++ dashboard/src/components/layout/sidebar.tsx
@@ -1,70 +1,8 @@
 import { Link, useLocation } from "@tanstack/react-router";
-import {
-  Activity,
-  AlertOctagon,
-  BarChart3,
-  Box,
-  CircuitBoard,
-  Cog,
-  ExternalLink as ExternalLinkIcon,
-  LayoutDashboard,
-  ListTree,
-  type LucideIcon,
-  ScrollText,
-  Server,
-  Settings2,
-  Skull,
-  Webhook as WebhookIcon,
-} from "lucide-react";
+import { AlertOctagon, ExternalLink as ExternalLinkIcon } from "lucide-react";
 import { useBranding, useExternalLinks } from "@/features/settings";
 import { cn } from "@/lib/cn";
-
-interface NavItem {
-  to: string;
-  label: string;
-  icon: LucideIcon;
-}
-
-interface NavGroup {
-  title: string;
-  items: NavItem[];
-}
-
-const NAV: NavGroup[] = [
-  {
-    title: "Monitoring",
-    items: [
-      { to: "/", label: "Overview", icon: LayoutDashboard },
-      { to: "/jobs", label: "Jobs", icon: ListTree },
-      { to: "/metrics", label: "Metrics", icon: BarChart3 },
-      { to: "/logs", label: "Logs", icon: ScrollText },
-    ],
-  },
-  {
-    title: "Infrastructure",
-    items: [
-      { to: "/queues", label: "Queues", icon: Box },
-      { to: "/workers", label: "Workers", icon: Server },
-      { to: "/resources", label: "Resources", icon: Activity },
-    ],
-  },
-  {
-    title: "Reliability",
-    items: [
-      { to: "/dead-letters", label: "Dead letters", icon: Skull },
-      { to: "/circuit-breakers", label: "Circuit breakers", icon: CircuitBoard },
-      { to: "/system", label: "System", icon: Settings2 },
-    ],
-  },
-  {
-    title: "Configuration",
-    items: [
-      { to: "/tasks", label: "Tasks", icon: ListTree },
-      { to: "/webhooks", label: "Webhooks", icon: WebhookIcon },
-      { to: "/settings", label: "Settings", icon: Cog },
-    ],
-  },
-];
+import { NAV } from "./nav-config";
 
 export function Sidebar() {
   const { pathname } = useLocation();
